+++ src/main/java/stage/sir/gestioncomptabilite/vo/DeclarationIsXml.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 @XmlRootElement(name = "DeclarationIS")
 public class DeclarationIsXml {
-    private long id;
+    private Long id;
     private String ref;
-    private double annee;
+    private Integer annee;
     private Double totalHTGain;
     private Double totalHTCharge;
     private Double totalHTDiff;
@@ -23,11 +23,11 @@ public class DeclarationIsXml {
     private List<Facture> facturesC;*/
     private List<Facture> factures;
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -39,11 +39,11 @@ public class DeclarationIsXml {
         this.ref = ref;
     }
 
-    public double getAnnee() {
+    public Integer getAnnee() {
         return annee;
     }
 
-    public void setAnnee(double annee) {
+    public void setAnnee(Integer annee) {
         this.annee = annee;
     }
